server: consolidate files
