+++ server/util.go
@@ -14,6 +14,17 @@ import (
 	"github.com/tendermint/tmlibs/log"
 )
 
+type Context struct {
+	Config *cfg.Config
+	Logger log.Logger
+}
+
+func NewContext(config *cfg.Config, logger log.Logger) *Context {
+	return &Context{config, logger}
+}
+
+//--------------------------------------------------------------------
+
 // PersistentPreRunEFn returns a PersistentPreRunE function for cobra
 // that initailizes the passed in context with a properly configured
 // logger and config objecy
